+++ CHANGES.rst
@@ -5,7 +5,8 @@ Changelog
 1.16.3 (unreleased)
 -------------------
 
-- Nothing changed yet.
+- Updated is_permitted function to check multiple permissions.
+  [chris-adam]
 
 
 1.16.2 (2025-02-28)